+++ src/lib/whitespace/nbsp.js
@@ -12,7 +12,7 @@ export function removeNbspBetweenMultiCharWords(string, locale) {
 
 
 export function addNbspAfterPreposition(string, locale) {
-	let pattern = "(^|[" + locale.space + "]|[^" + locale.allChars + "])([" + locale.allChars + "])([" + locale.space + "])"
+	let pattern = "(^|[" + locale.space + "]|[^" + locale.allChars + locale.apostrophe + "])([" + locale.allChars + "])([" + locale.space + "])"
 	let re = new RegExp(pattern, "g");
 	let replacement = "$1$2" + locale.nbsp;
 	string = string.replace(re, replacement);