+++ src/test/punctuation/hyphen.test.js
@@ -1,26 +1,8 @@
 import {fixHyphen,
-				replaceDashWithHyphen,
 				fixSpaceAroundHyphen} from "../../lib/punctuation/hyphen";
 import assert from 'assert';
 import Locale from "../../locale/locale";
 
-
-describe('Replace dashes with hyphen, where applicable\n', () => {
-	let testCase = {
-		"two—year—old child": "two-year-old child",
-		"two–year–old child": "two-year-old child",
-		"zeleno–žltá": "zeleno-žltá",
-		};
-
-	Object.keys(testCase).forEach((key) => {
-		it("unit tests", () => {
-			assert.equal(replaceDashWithHyphen(key, new Locale("en-us")), testCase[key]);
-		});
-		it("module tests", () => {
-			assert.equal(fixHyphen(key, new Locale("en-us")), testCase[key]);
-		});
-	});
-});
 
 
 describe('Fix spaces around hyphen\n', () => {